feat(engine): move module declarations outside of plugin file
