+++ src/engine/Engine.hpp
@@ -23,25 +23,10 @@
 #include "Velocity.hpp"
 #include "Window.hpp"
 #include "ecs.hpp"
+#include "engine.hpp"
 #include "glWindow.hpp"
 
 namespace cevy::engine {
-
-#ifdef DEBUG
-struct DebugWindow {
-  bool open;
-};
-#endif
-
-class StartupRenderStage : public cevy::ecs::core_stage::after<cevy::ecs::core_stage::PreStartup> {
-};
-class PreStartupRenderStage : public cevy::ecs::core_stage::before<StartupRenderStage> {};
-class PostStartupRenderStage : public cevy::ecs::core_stage::after<StartupRenderStage> {};
-
-class RenderStage : public cevy::ecs::core_stage::after<cevy::ecs::core_stage::PostUpdate> {};
-class PreRenderStage : public cevy::ecs::core_stage::before<RenderStage> {};
-class PostRenderStage : public cevy::ecs::core_stage::after<RenderStage> {};
-
 template <template <typename T> typename Windower = glWindow, typename Renderer = ForwardRenderer>
 class Engine : public cevy::ecs::Plugin {
   public:
